pyproject.toml
[project]
name = "a2a-inspector"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.6",
    "httpx>=0.28.1",
    "httpx-sse>=0.4.0",
    "jwcrypto>=1.5.6",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "sse-starlette>=2.2.1",
    "starlette>=0.46.1",
    "typing-extensions>=4.12.2",
    "uvicorn>=0.34.0",
    "Flask",
    "flask_cors",
]
[tool.uv.workspace]
members = [
    "inspector",
]
